fix(clippy): apply 7 of 11 mechanical residuals from pre-release triage
Apply the non-collision clippy warnings from the 2026-06-13 pre-release
code review (cargo clippy --lib, default features). All are pure
one-liner rewrites with identical behaviour.

is_multiple_of (3 of 6):
- src/int/algos/div/div_knuth_u128_limb.rs:139
- src/int/policy/div_rem.rs:238
- src/int/policy/mul.rs:121

div_ceil (2 of 2):
- src/int/algos/mul/mul_toom3.rs:78
- src/int/algos/mul/mul_toom3.rs:194

unnecessary same-type casts (2 of 2):
- src/int/algos/cube/cube_fused_comba.rs:82
- src/int/algos/hypot/hypot_u128_fast.rs:227

Skipped-for-collision (coordinator sweeps after sibling lands):
- src/algos/support/newton_reciprocal.rs:230,661 (is_multiple_of)
- src/int/types/compute_limbs.rs:227 (is_multiple_of)
- src/algos/support/wide_trig_core.rs:1667 (too_many_arguments allow)

+++ src/int/algos/cube/cube_fused_comba.rs
@@ -79,7 +79,7 @@ pub(crate) const fn cube_fused_comba<const N: usize>(x: Uint<N>) -> Uint<N> {
                 // p = x_i · x_j · x_k, up to 192 bits, held as (p_lo:u128, p_hi:u64).
                 let ab = xi * xj; // ≤ (2^64−1)² < 2^128
                 let ab_lo = ab as u64 as u128;
-                let ab_hi = (ab >> 64) as u128;
+                let ab_hi = ab >> 64;
                 // (ab_lo + ab_hi·2^64) · xk
                 let t0 = ab_lo * xk; // ≤ 2^128
                 let t1 = ab_hi * xk; // ≤ 2^128
